docs: merge results start_here.py with aggregator into single tutorial
Combine guides/results/start_here.py (simple JSON/FITS loading) with
guides/results/aggregator/start_here.py (full aggregator walkthrough)
into one entry-point tutorial at guides/results/start_here.py. Simple
loading is presented first with .exists() guards so the placeholder
<unique_hash> path doesn't crash, then the aggregator section runs a
real fit and walks the deeper API. Cross-references inside the
aggregator section note when each result is also reachable via the
simple-loading API.

Mirror the existing autogalaxy test_mode_was_on / n_like_max=300
conditional into autolens_workspace's start_here.py for parity, so a
manual PYAUTO_TEST_MODE=1 run produces 300 samples instead of being
short-circuited to a mock result. Verified end-to-end: search.json
records n_like_max=300 in both workspaces, samples_info.json reports
total_accepted_samples=300, downstream aggregator/samples.py loads
the result cleanly.

Delete the now-redundant guides/results/aggregator/start_here.py and
collapse the dual pointer in scripts/imaging/modeling.py into a single
pointer noting both APIs are covered.
