+++ .script_sizes.json
@@ -37,14 +37,13 @@
   "scripts/guides/results/aggregator/queries.py": 6008,
   "scripts/guides/results/aggregator/samples.py": 20061,
   "scripts/guides/results/aggregator/samples_via_aggregator.py": 19322,
-  "scripts/guides/results/aggregator/start_here.py": 18643,
   "scripts/guides/results/database/__init__.py": 0,
   "scripts/guides/results/database/simulators/__init__.py": 0,
   "scripts/guides/results/database/simulators/light_sersic_exp__0.py": 6730,
   "scripts/guides/results/database/simulators/light_sersic_exp__1.py": 6729,
   "scripts/guides/results/database/simulators/light_sersic_exp__2.py": 6728,
   "scripts/guides/results/database/start_here.py": 14256,
-  "scripts/guides/results/start_here.py": 5542,
+  "scripts/guides/results/start_here.py": 25225,
   "scripts/guides/results/workflow/__init__.py": 0,
   "scripts/guides/results/workflow/csv_make.py": 14185,
   "scripts/guides/results/workflow/fits_make.py": 15320,
@@ -59,11 +58,11 @@
   "scripts/imaging/data_preparation/examples/data.py": 9289,
   "scripts/imaging/data_preparation/examples/noise_map.py": 5499,
   "scripts/imaging/data_preparation/examples/optional/__init__.py": 0,
-  "scripts/imaging/data_preparation/examples/optional/extra_galaxies_centres.py": 4553,
+  "scripts/imaging/data_preparation/examples/optional/extra_galaxies_centres.py": 4593,
   "scripts/imaging/data_preparation/examples/optional/info.py": 2844,
   "scripts/imaging/data_preparation/examples/optional/light_centre.py": 3311,
   "scripts/imaging/data_preparation/examples/optional/mask.py": 5094,
-  "scripts/imaging/data_preparation/examples/optional/mask_extra_galaxies.py": 4740,
+  "scripts/imaging/data_preparation/examples/optional/mask_extra_galaxies.py": 5533,
   "scripts/imaging/data_preparation/examples/psf.py": 6845,
   "scripts/imaging/data_preparation/gui/__init__.py": 0,
   "scripts/imaging/data_preparation/gui/extra_galaxies_centres.py": 3524,
@@ -75,8 +74,8 @@
   "scripts/imaging/data_preparation/start_here.py": 15373,
   "scripts/imaging/features/__init__.py": 0,
   "scripts/imaging/features/extra_galaxies/__init__.py": 0,
-  "scripts/imaging/features/extra_galaxies/modeling.py": 17529,
-  "scripts/imaging/features/extra_galaxies/simulator.py": 7137,
+  "scripts/imaging/features/extra_galaxies/modeling.py": 17244,
+  "scripts/imaging/features/extra_galaxies/simulator.py": 8906,
   "scripts/imaging/features/linear_light_profiles/__init__.py": 0,
   "scripts/imaging/features/linear_light_profiles/fit.py": 8180,
   "scripts/imaging/features/linear_light_profiles/likelihood_function.py": 23944,
@@ -90,9 +89,9 @@
   "scripts/imaging/features/operated_light_profile/modeling.py": 9082,
   "scripts/imaging/features/operated_light_profile/simulator.py": 4803,
   "scripts/imaging/features/pixelization/__init__.py": 0,
-  "scripts/imaging/features/pixelization/fit.py": 17111,
+  "scripts/imaging/features/pixelization/fit.py": 16827,
   "scripts/imaging/features/pixelization/likelihood_function.py": 35526,
-  "scripts/imaging/features/pixelization/modeling.py": 20296,
+  "scripts/imaging/features/pixelization/modeling.py": 20012,
   "scripts/imaging/features/pixelization/source_science.py": 6288,
   "scripts/imaging/features/shapelets/__init__.py": 0,
   "scripts/imaging/features/shapelets/fit.py": 15693,
@@ -104,11 +103,11 @@
   "scripts/imaging/features/sky_background/simulator.py": 4934,
   "scripts/imaging/fit.py": 15512,
   "scripts/imaging/likelihood_function.py": 13685,
-  "scripts/imaging/modeling.py": 23267,
+  "scripts/imaging/modeling.py": 23205,
   "scripts/imaging/simulator.py": 6446,
   "scripts/imaging/simulator_sample.py": 6356,
-  "scripts/imaging/simulator_sersic.py": 4030,
-  "scripts/imaging/start_here.py": 20211,
+  "scripts/imaging/simulator_sersic.py": 4932,
+  "scripts/imaging/start_here.py": 19926,
   "scripts/interferometer/__init__.py": 0,
   "scripts/interferometer/casa_reduction.py": 7259,
   "scripts/interferometer/data_preparation.py": 12068,
